kvutils: Use Raw#size when computing cache weight.

+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/Raw.scala
@@ -8,7 +8,7 @@ import com.google.protobuf.ByteString
 sealed trait Raw {
   def bytes: ByteString
 
-  final def size: Int = bytes.size
+  final def size: Long = bytes.size.toLong
 }
 
 object Raw {
